Several things bundled.
Mostly this is a structural commit. After poking at hubot, ripple-ui and cordova a bit to see how they handle project structure etc I'm settling on something like a style and folder structure.

eventually murdules will be included based on the murdules.json file, for now just the echo murdule is hardcoded.

the correct message text (based on my very limited sample set of testing through gchat) is passed to the murdule and it is responding with text, but right now the application itself still isn't doing anything with that.

next step is to put the output of the murdule into the stanza itself, which means more xmpp digging.